+++ attune/_cli.py
@@ -25,28 +25,23 @@ def catalog():
 
 @main.command(name="history", help="show the change history of an instrument")
 @click.argument("instrument", nargs=1)
-@click.option("-n", default=10, help="number of change records to list")
-# TODO: add a --date option (perhaps just date) to anchor the history list to
-# TODO: add a --direction option to search forward or backward from date
-def history(instrument, n=10):
-    title_string = f"{instrument}, from latest to earliest"
-    print(title_string + "-" * (80 - len(instrument)))
-    current = store.load(instrument)
-    for i in range(n):
-        try:
-            print(
-                "{0:4} {1}{2} at {3}".format(
-                    -i,
-                    current.transition.type,
-                    "." * (20 - len(current.transition.type)),
-                    str(current.load),
-                )
-            )
-            current = store.undo(current)
-        except ValueError:  # reached end of history
-            print("<end of history>")
-            break
-
+@click.option("-n", default=10, help="number of records to list (default is 10)")
+@click.option("--start", "-s", default="now", help="date to start history (default is now)")
+@click.option(
+    "--forward",
+    "-f",
+    is_flag=True,
+    show_default=True,
+    default=False,
+    help="when specified, history will search forwards in time"
+)
+def history(instrument, n=10, start="now", forward=False):
+    try:
+        root = store.load(instrument, start)
+    except ValueError as e:
+        print(e)
+        return
+    root.print_history(n, not forward)
 
 if __name__ == "__main__":
     main()