+++ attune/__init__.py
@@ -19,3 +19,5 @@ from ._tune import *
 from ._tune_test import *
 from ._update_merge import *
 from .io import *
+
+from . import _cli