+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Added
+- command line interface for simple queries of instruments
+- tunes now have a `monotonic` property
+- instruments now have a `print_tree` method
+
 ### Fixed
 - Inserted WrightTools import into _map and _offset